A few changes to the distance calculation and LCD output
Using the built in distance calculation and Cardinality (N/E/S/W) support
to populate the LCD display.

A few issues with float to string convertion, the dtostrf() function
is included but commented out as it adds 2k to the build size.

I've removed the conditional power saving stuff as I plan to make this
as low powered as possible.

Target is now hardcoded to The Bell Inn, Bath, because why not.